backend/requirements.txt
asgiref==3.5.0
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
